azure/blob_storage/requirements.txt
python-dotenv
azure-identity
azure-storage-blob